tables: repurpose inverse quant table

+++ schroedinger/schrotables.h
@@ -12,7 +12,7 @@ SCHRO_BEGIN_DECLS
 extern const uint32_t schro_table_offset_3_8[61];
 extern const uint32_t schro_table_offset_1_2[61];
 extern const uint32_t schro_table_quant[61];
-extern const uint32_t schro_table_inverse_quant[61];
+extern const uint16_t schro_table_inverse_quant[61];
 extern const uint16_t schro_table_division_factor[257];
 extern const double schro_table_error_hist_shift3_1_2[60][104];
 extern const double schro_table_onebits_hist_shift3_1_2[60][104];
